pitron practice

+++ week 01 -- Introduction/Module --02 STL vector/Vector Capacity function 02.cpp
@@ -2,6 +2,6 @@
 using namespace std;
 int main()
 {
-vector <long long int> v;
-cout << v.max_size();
+    vector <int > v;
+    cout << v.max_size();
 }
